Made new cursor trail default and removed previous cursors
