+++ components/cursor/cursorTrail.tsx
@@ -1,69 +1,133 @@
-import React, { useEffect, useState, useRef } from "react";
+import React, { useEffect, useState, useRef, CSSProperties } from "react";
+import Image from "next/image";
 
 type Cursor = {
-  top: number;
-  left: number;
+  Y: number;
+  X: number;
   key: number;
 };
 
-function RenderCursorTrail() {
+const CURSOR_MIN_DISTANCE = 100; // Min distance between trail points to add a new one
+const CURSOR_TIMEOUT_MS = 100; // Time (ms) for a trail point to live and for GIF to travel to it
+
+function CursorTrailWithGif() {
   const [cursors, setCursors] = useState<Cursor[]>([]);
-  const lastCursor = useRef<Cursor | null>(null);
+  const [gifStyle, setGifStyle] = useState<CSSProperties>({
+    position: "absolute",
+    top: "-100px", // Start off-screen
+    left: "-100px",
+    transform: "translate(-50%, -50%) scaleX(1)",
+    pointerEvents: "none",
+    // Initial transition set, will be updated dynamically
+    transition: `top ${CURSOR_TIMEOUT_MS / 1000}s linear, left ${
+      CURSOR_TIMEOUT_MS / 1000
+    }s linear`,
+    willChange: "top, left", // Hint for performance
+  });
+  const lastCursorPos = useRef<Cursor | null>(null);
+  const currentGifTargetKey = useRef<number | null>(null); // Track the key of the target cursor
+  const previousGifTargetPos = useRef<{ Y: number; left: number } | null>(null); // Track previous target for flipping logic
+  const timeoutIds = useRef(new Map<number, NodeJS.Timeout>());
 
-  const cursorMinDistance = 100; // Change this value to adjust the distance between cursors
-  const cursorTimeOut = 500; // Change this value to adjust how long each div stays on the screen
-
+  // Effect to handle mouse movement and add trail points
   useEffect(() => {
-    const timeoutIds = new Map<number, NodeJS.Timeout>();
-
-    const moveCursor = (e: MouseEvent) => {
+    const moveMouse = (e: MouseEvent) => {
       const potentialNewCursor: Cursor = {
-        top: e.clientY,
-        left: e.clientX,
+        Y: e.clientY,
+        X: e.clientX,
         key: Date.now(),
       };
 
-      // Calculate the distance between the new cursor and the last cursor
-      const distance = lastCursor.current
+      // Calculate distance from the last *added* trail point
+      const distance = lastCursorPos.current
         ? Math.sqrt(
-            Math.pow(potentialNewCursor.left - lastCursor.current.left, 2) +
-              Math.pow(potentialNewCursor.top - lastCursor.current.top, 2),
+            Math.pow(potentialNewCursor.X - lastCursorPos.current.X, 2) +
+              Math.pow(potentialNewCursor.Y - lastCursorPos.current.Y, 2),
           )
         : Infinity;
 
-      // Only add the new cursor if the distance is greater than the threshold
-      if (distance > cursorMinDistance) {
-        const keyToRemove = potentialNewCursor.key;
+      // Only add a new point if the mouse moved sufficiently
+      if (distance > CURSOR_MIN_DISTANCE) {
+        const newKey = potentialNewCursor.key;
 
+        // Add the new cursor point to the state
         setCursors((prevCursors) => {
-          // Calculate timeout based on the position (index + 1) in the trail
-          const newCursorIndex = prevCursors.length + 1;
-          const timeoutDuration = newCursorIndex * cursorTimeOut;
+          const newCursors = [...prevCursors, potentialNewCursor];
+          const timeoutDuration = newCursors.length * CURSOR_TIMEOUT_MS; // Stagger removal
 
+          // Set a timeout to remove this specific cursor point after its calculated duration
           const timeoutId = setTimeout(() => {
             setCursors((currentCursors) =>
-              currentCursors.filter((cursor) => cursor.key !== keyToRemove),
+              currentCursors.filter((cursor) => cursor.key !== newKey),
             );
-            timeoutIds.delete(keyToRemove);
+            timeoutIds.current.delete(newKey);
           }, timeoutDuration);
 
-          timeoutIds.set(keyToRemove, timeoutId);
-
-          return [...prevCursors, potentialNewCursor];
+          timeoutIds.current.set(newKey, timeoutId);
+          return newCursors;
         });
-
-        lastCursor.current = potentialNewCursor;
+        lastCursorPos.current = potentialNewCursor; // Update last added position
       }
     };
 
-    window.addEventListener("mousemove", moveCursor);
+    window.addEventListener("mousemove", moveMouse);
 
+    // Cleanup function
     return () => {
-      window.removeEventListener("mousemove", moveCursor);
-      timeoutIds.forEach((timeoutId) => clearTimeout(timeoutId));
+      window.removeEventListener("mousemove", moveMouse);
+      // Clear any remaining timeouts when the component unmounts
+      timeoutIds.current.forEach((timeoutId) => clearTimeout(timeoutId));
+      timeoutIds.current.clear();
     };
-    // Add dependencies to re-run the effect if these values change
-  }, [cursorMinDistance, cursorTimeOut]);
+    // Only CURSOR_MIN_DISTANCE and CURSOR_TIMEOUT_MS are dependencies here
+    // Cursors state is managed internally via setCursors callback
+  }, [CURSOR_MIN_DISTANCE, CURSOR_TIMEOUT_MS]);
+
+  // Effect to update GIF target and style when the trail changes
+  useEffect(() => {
+    if (cursors.length > 0) {
+      const nextTarget = cursors[0]; // The oldest cursor is the next target
+
+      // Only update if the target has actually changed
+      if (nextTarget.key !== currentGifTargetKey.current) {
+        // Determine flip direction based on previous target
+        // Use current GIF position if previous target isn't set yet (first move)
+        const currentLeft =
+          previousGifTargetPos.current?.left ??
+          parseFloat(gifStyle.left as string) ??
+          nextTarget.X;
+        const isFlipped = nextTarget.X < currentLeft;
+
+        setGifStyle((prevStyle) => ({
+          ...prevStyle,
+          top: `${nextTarget.Y}px`,
+          left: `${nextTarget.X}px`,
+          transform: `translate(-50%, -50%) scaleX(${isFlipped ? -1 : 1})`,
+          // Ensure transition matches the time until the *next* point expires
+          transition: `top ${CURSOR_TIMEOUT_MS / 1000}s linear, left ${
+            CURSOR_TIMEOUT_MS / 1000
+          }s linear`,
+        }));
+
+        currentGifTargetKey.current = nextTarget.key;
+        previousGifTargetPos.current = {
+          Y: nextTarget.Y,
+          left: nextTarget.X,
+        };
+      }
+    } else {
+      // No cursors, reset target key and potentially hide GIF or stop transition
+      currentGifTargetKey.current = null;
+      previousGifTargetPos.current = null; // Reset previous position too
+      // Optional: Move GIF off-screen immediately when idle and stop transition
+      // setGifStyle(prevStyle => ({
+      //   ...prevStyle,
+      //   top: '-100px',
+      //   left: '-100px',
+      //   transition: 'none' // Stop transition when idle
+      // }));
+    }
+  }, [cursors, CURSOR_TIMEOUT_MS, gifStyle.left]); // Include gifStyle.left to recalculate flip correctly on first move
 
   return (
     <div>
@@ -71,16 +135,26 @@ function RenderCursorTrail() {
         <div
           key={cursor.key}
           className="pointer-events-none invisible absolute h-5 w-5 rounded-full bg-black lg:visible dark:bg-white"
-          style={{ top: cursor.top, left: cursor.left, position: "absolute" }}
+          style={{
+            top: `${cursor.Y}px`,
+            left: `${cursor.X}px`,
+            position: "absolute",
+            transform: "translate(-50%, -50%)", // Center the dot on the coordinate
+          }}
         />
       ))}
+
+      <Image
+        src="/pacman.gif"
+        width={40}
+        height={40}
+        alt="Following GIF"
+        className="invisible lg:visible"
+        style={gifStyle} // Apply the calculated style
+        unoptimized // NextJS cannot optimize Gifs
+      />
     </div>
   );
 }
-export default function customCursorTrail() {
-  return (
-    <div>
-      <RenderCursorTrail />
-    </div>
-  );
-}
+
+export default CursorTrailWithGif;
